+++ attributes/default.rb
@@ -1,12 +1,4 @@
-default['gflags']['install_type'] = value_for_platform(
-  %w{centos fedora redhat} => {
-    "default" => "package"
-  },
-  %w{ubuntu} => {
-    "12.04" => "archive",
-    "default" => "package"
-  }
-)
+default['gflags']['install_type'] = nil
 
 # Source attributes
 default['gflags']['archive']['install_dir'] = "/usr/local"