+++ recipes/package.rb
@@ -1,2 +1,0 @@
-include_recipe "gflags::package_cpp"
-include_recipe "gflags::package_python"